backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "zeocax", email = "zeocax@zeocax.com"},
]
dependencies = [
    "fastapi[standard]>=0.115.12",
    "sqlalchemy>=2.0.41",
    "pydantic[email]>=2.11.4",
    "pydantic-extra-types>=2.10.4",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "pyjwt>=2.10.1",
]
requires-python = ">3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
